Revert "Feature/onboarding splash screens dependent on pr #01"
