--- ivoluntia/app/(app)/_layout.tsx
@@ -1,19 +1,0 @@
-import { Stack } from 'expo-router';
-
-export default function AppLayout() {
-  return (
-    <Stack
-      screenOptions={{
-        headerStyle: {
-          backgroundColor: '#f4511e',
-        },
-        headerTintColor: '#fff',
-        headerTitleStyle: {
-          fontWeight: 'bold',
-        },
-      }}>
-      <Stack.Screen name="index" options={{ title: 'Home' }} />
-      {/* Define other main app screens here */}
-    </Stack>
-  );
-}
