+++ ivoluntia/app/index.tsx
@@ -1,0 +1,23 @@
+import { View, Text, StyleSheet } from "react-native";
+
+export default function Onboarding() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.text}>Hey Trustees</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor: "#fff",
+  },
+  text: {
+    fontSize: 28,
+    fontWeight: "bold",
+    color: "#333",
+  },
+});
